refactor: refactor all page to use layout to reduce redundant code
