+++ src/app/(auth-page)/(login-register)/login/page.tsx
@@ -1,0 +1,5 @@
+import { LoginForm } from '@/components/auth/(login-register)/login/login-form';
+
+export default function LoginPage() {
+  return <LoginForm />;
+}
