+++ src/app/(auth-page)/(login-register)/register/page.tsx
@@ -1,10 +1,5 @@
-import { Layout } from '@/components/auth/(login-register)/layout';
 import { RegisterForm } from '@/components/auth/(login-register)/register/register-form';
 
 export default function RegisterPage() {
-  return (
-    <Layout>
-      <RegisterForm />
-    </Layout>
-  );
+  return <RegisterForm />;
 }
