--- src/app/(auth-page)/login/page.tsx
@@ -1,10 +1,0 @@
-import { Layout } from '@/components/auth/(login-register)/layout';
-import { LoginForm } from '@/components/auth/(login-register)/login/login-form';
-
-export default function LoginPage() {
-  return (
-    <Layout>
-      <LoginForm />
-    </Layout>
-  );
-}
