+++ src/app/(public-page)/catalog/page.tsx
@@ -1,0 +1,5 @@
+﻿import { CatalogMain } from '@/components/catalog/main/catalog-main';
+
+export default function CatalogPage() {
+  return <CatalogMain />;
+}
