+++ src/app/(public-page)/page.tsx
@@ -1,0 +1,26 @@
+import {
+  EncyclopediaSection,
+  FeaturedCards,
+  HeroSection,
+  IslandCards,
+  ProductCatalog,
+} from '@/components/landing-page';
+
+export default function HomePage() {
+  return (
+    <main className="pb-16">
+      <section className="mx-auto w-full max-w-[1320px] px-4 pt-8 md:px-6 lg:px-8">
+        <div className="grid animate-in gap-4 fade-in duration-700 lg:grid-cols-[minmax(0,1fr)_280px]">
+          <HeroSection />
+          <FeaturedCards />
+        </div>
+      </section>
+
+      <ProductCatalog />
+
+      <EncyclopediaSection />
+
+      <IslandCards />
+    </main>
+  );
+}
